requirements.txt
Django==3.1
djangorestframework==3.12
psycopg2-binary==2.8.6
Pillow==7.1
python-decouple
django-filter
drf-yasg
transliterate==1.10.2
djangorestframework_simplejwt
celery==5.1.1
BeautifulSoup4
requests
lxml